requirements.txt
pandas
numpy
joblib
matplotlib
scikit-learn
tensorflow-cpu
streamlit